--- packages/react-examples/src/react-next/Coachmark/docs/CoachmarkBestPractices.md
@@ -1,6 +1,0 @@
-## Layout
-
-- Only one coach mark and teaching bubble combo should be displayed at a time.
-- Coach marks can be standalone or sequential. Sequential coach marks should be used sparingly to walk through complex multistep interactions. It’s recommended that a sequence of coach marks doesn’t exceed three steps.
-- Coach marks are designed to only hold teaching bubbles.
-- Coach mark size, color, and animation shouldn’t be altered.
