--- packages/react-examples/src/react-next/Coachmark/docs/CoachmarkOverview.md
@@ -1,1 +1,0 @@
-Coach marks (`Coachmark`) are used to draw a person’s attention to parts of the UI and increase engagement with those elements. A teaching bubble appears on hover or selection of the coach mark.
