+++ packages/react-examples/src/react/Coachmark/Coachmark.Basic.Example.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react';
 import {
   Coachmark,
-  DefaultButton,
   DirectionalHint,
   Dropdown,
-  IButtonProps,
   IDropdownOption,
   TeachingBubbleContent,
   mergeStyleSets,
 } from '@fluentui/react';
+import { DefaultButton, IButtonProps } from '@fluentui/react-next/lib/compat/Button';
 import { useBoolean } from '@uifabric/react-hooks';
 
 const classNames = mergeStyleSets({
