+++ packages/react-examples/src/react/Coachmark/Coachmark.doc.tsx
@@ -7,7 +7,7 @@ const CoachmarkBasicExampleCode = require('!raw-loader!@fluentui/react-examples/
 export const CoachmarkPageProps: IDocPageProps = {
   title: 'Coachmark',
   componentName: 'Coachmark',
-  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react/src/components/Coachmark',
+  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react-internal/src/components/Coachmark',
   examples: [
     {
       title: 'Coachmark Basic',
