+++ packages/react-internal/src/components/Coachmark/Coachmark.base.tsx
@@ -7,24 +7,23 @@ import {
   getDocument,
   IRectangle,
   KeyCodes,
-  shallowCompare,
   getRTL,
-  warnDeprecations,
   EventGroup,
-  Async,
-  initializeComponentRef,
+  getPropsWithDefaults,
 } from '../../Utilities';
-import { IPositionedData, RectangleEdge, getOppositeEdge } from '../../utilities/positioning';
+import { IPositionedData, RectangleEdge, getOppositeEdge } from '../../Positioning';
 
 // Component Dependencies
-import { PositioningContainer, IPositioningContainer } from './PositioningContainer/index';
+import { PositioningContainer } from './PositioningContainer/index';
 import { Beak, BEAK_HEIGHT, BEAK_WIDTH } from './Beak/Beak';
 import { DirectionalHint } from '../../common/DirectionalHint';
 
 // Coachmark
-import { ICoachmark, ICoachmarkProps, ICoachmarkStyles, ICoachmarkStyleProps } from './Coachmark.types';
+import { ICoachmarkProps, ICoachmarkStyles, ICoachmarkStyleProps } from './Coachmark.types';
 import { COACHMARK_HEIGHT, COACHMARK_WIDTH } from './Coachmark.styles';
 import { FocusTrapZone } from '../../FocusTrapZone';
+import { useAsync, useOnEvent, useSetTimeout, useWarnings } from '@uifabric/react-hooks';
+import { IBeakProps } from './Beak/Beak.types';
 
 const getClassNames = classNamesFunction<ICoachmarkStyleProps, ICoachmarkStyles>();
 
@@ -34,500 +33,128 @@ export const COACHMARK_ATTRIBUTE_NAME = 'data-coachmarkid';
  * An interface for the cached dimensions of entity inner host.
  */
 export interface IEntityRect {
-  width: number;
-  height: number;
+  width?: number;
+  height?: number;
 }
 
-export interface ICoachmarkState {
-  /**
-   * Is the Coachmark currently collapsed into
-   * a tear drop shape
-   */
-  isCollapsed: boolean;
+type BeakPosition = Pick<IBeakProps, 'left' | 'top' | 'right' | 'bottom' | 'direction'>;
 
-  /**
-   * Enables/Disables the beacon that radiates
-   * from the center of the coachmark.
-   */
-  isBeaconAnimating: boolean;
+const DEFAULT_PROPS: Partial<ICoachmarkProps> = {
+  isCollapsed: true,
+  mouseProximityOffset: 10,
+  delayBeforeMouseOpen: 3600, // The approximate time the coachmark shows up
+  delayBeforeCoachmarkAnimation: 0,
+  isPositionForced: true,
+  positioningContainerProps: {
+    directionalHint: DirectionalHint.bottomAutoEdge,
+  },
+};
 
-  /**
-   * Is the teaching bubble currently retreiving the
-   * original dimensions of the hosted entity.
-   */
-  isMeasuring: boolean;
+function useCollapsedState(props: ICoachmarkProps, entityInnerHostElementRef: React.RefObject<HTMLDivElement>) {
+  const { isCollapsed: propsIsCollapsed, onAnimationOpenStart, onAnimationOpenEnd } = props;
 
-  /**
-   * Is the Coachmark done measuring the hosted entity
-   */
-  isMeasured: boolean;
+  /** Is the Coachmark currently collapsed into a tear drop shape */
+  const [isCollapsed, setIsCollapsed] = React.useState<boolean>(!!propsIsCollapsed);
+  const { setTimeout } = useSetTimeout();
 
-  /**
-   * Cached width and height of _entityInnerHostElement
-   */
-  entityInnerHostRect: IEntityRect;
+  // Rather than pushing out logic elsewhere to prevent openCoachmark from being called repeatedly,
+  // we'll track it here and only invoke the logic once. We do this with a ref, rather than just the state,
+  // because the openCoachmark callback can be captured in scope for an effect
+  const hasCoachmarkBeenOpened = React.useRef(!isCollapsed);
 
-  /**
-   * Is the mouse in proximity of the default target element
-   */
-  isMouseInProximity: boolean;
+  const openCoachmark = React.useCallback(() => {
+    if (!hasCoachmarkBeenOpened.current) {
+      setIsCollapsed(false);
 
-  /**
-   * The left position of the beak
-   */
-  beakLeft?: string;
+      onAnimationOpenStart?.();
 
-  /**
-   * The right position of the beak
-   */
-  beakTop?: string;
+      entityInnerHostElementRef.current?.addEventListener?.('transitionend', (): void => {
+        // Need setTimeout to trigger narrator
+        setTimeout(() => {
+          if (entityInnerHostElementRef.current) {
+            focusFirstChild(entityInnerHostElementRef.current);
+          }
+        }, 1000);
 
-  /**
-   * The right position of the beak
-   */
-  beakRight?: string;
+        onAnimationOpenEnd?.();
+      });
+      hasCoachmarkBeenOpened.current = true;
+    }
+  }, [entityInnerHostElementRef, onAnimationOpenEnd, onAnimationOpenStart, setTimeout]);
 
-  /**
-   * The bottom position of the beak
-   */
-  beakBottom?: string;
+  React.useEffect(() => {
+    if (!propsIsCollapsed) {
+      openCoachmark();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- should only run if isCollapsed changes
+  }, [propsIsCollapsed]);
+
+  return [isCollapsed, openCoachmark] as const;
+}
+
+function usePositionedData() {
+  const async = useAsync();
 
   /**
    * Alignment edge of callout in relation to target
    */
-  targetAlignment?: RectangleEdge;
-
+  const [targetAlignment, setTargetAlignment] = React.useState<RectangleEdge | undefined>();
   /**
    * Position of Coachmark/TeachingBubble in relation to target
    */
-  targetPosition?: RectangleEdge;
+  const [targetPosition, setTargetPosition] = React.useState<RectangleEdge | undefined>();
 
-  /**
-   * Transform origin of teaching bubble callout
-   */
-  transformOrigin?: string;
+  const onPositioned = ({ alignmentEdge, targetEdge }: IPositionedData) =>
+    async.requestAnimationFrame(() => {
+      setTargetAlignment(alignmentEdge);
+      setTargetPosition(targetEdge);
+    });
 
-  /**
-   * ARIA alert text to read aloud with Narrator once the Coachmark is mounted
-   */
-  alertText?: string;
+  return [targetAlignment, targetPosition, onPositioned] as const;
 }
 
-const COMPONENT_NAME = 'Coachmark';
+function useBeakPosition(
+  props: ICoachmarkProps,
+  targetAlignment: RectangleEdge | undefined,
+  targetPosition: RectangleEdge | undefined,
+) {
+  const isRTL = getRTL(props.theme);
 
-export class CoachmarkBase extends React.Component<ICoachmarkProps, ICoachmarkState> implements ICoachmark {
-  public static defaultProps: Partial<ICoachmarkProps> = {
-    isCollapsed: true,
-    mouseProximityOffset: 10,
-    delayBeforeMouseOpen: 3600, // The approximate time the coachmark shows up
-    delayBeforeCoachmarkAnimation: 0,
-    isPositionForced: true,
-    positioningContainerProps: {
-      directionalHint: DirectionalHint.bottomAutoEdge,
-    },
-  };
+  return React.useMemo(() => {
+    const beakDirection = targetPosition === undefined ? RectangleEdge.bottom : getOppositeEdge(targetPosition);
 
-  private _async: Async;
-  private _events: EventGroup;
+    const beakPosition: BeakPosition = { direction: beakDirection };
 
-  /**
-   * The cached HTMLElement reference to the Entity Inner Host
-   * element.
-   */
-  private _entityHost = React.createRef<HTMLDivElement>();
-  private _entityInnerHostElement = React.createRef<HTMLDivElement>();
-  private _translateAnimationContainer = React.createRef<HTMLDivElement>();
-  private _ariaAlertContainer = React.createRef<HTMLDivElement>();
-  private _childrenContainer = React.createRef<HTMLDivElement>();
-  private _positioningContainer = React.createRef<IPositioningContainer>();
+    let transformOriginX: string;
+    let transformOriginY: string;
 
-  /**
-   * The target element the mouse would be in
-   * proximity to
-   */
-  private _targetElementRect: ClientRect;
-
-  constructor(props: ICoachmarkProps) {
-    super(props);
-
-    this._async = new Async(this);
-    this._events = new EventGroup(this);
-    initializeComponentRef(this);
-
-    warnDeprecations(COMPONENT_NAME, props, {
-      teachingBubbleRef: undefined,
-      collapsed: 'isCollapsed',
-      beakWidth: undefined,
-      beakHeight: undefined,
-      width: undefined,
-      height: undefined,
-    });
-
-    // Set defaults for state
-    this.state = {
-      isCollapsed: props.isCollapsed!,
-      isBeaconAnimating: true,
-      isMeasuring: true,
-      entityInnerHostRect: {
-        width: 0,
-        height: 0,
-      },
-      isMouseInProximity: false,
-      isMeasured: false,
-    };
-  }
-
-  private get _beakDirection(): RectangleEdge {
-    const { targetPosition } = this.state;
-    if (targetPosition === undefined) {
-      return RectangleEdge.bottom;
-    }
-
-    return getOppositeEdge(targetPosition);
-  }
-
-  public render(): JSX.Element {
-    const {
-      beaconColorOne,
-      beaconColorTwo,
-      children,
-      target,
-      color,
-      positioningContainerProps,
-      ariaDescribedBy,
-      ariaDescribedByText,
-      ariaLabelledBy,
-      ariaLabelledByText,
-      ariaAlertText,
-      delayBeforeCoachmarkAnimation,
-      styles,
-      theme,
-      className,
-      persistentBeak,
-    } = this.props;
-
-    const {
-      beakLeft,
-      beakTop,
-      beakRight,
-      beakBottom,
-      isCollapsed,
-      isBeaconAnimating,
-      isMeasuring,
-      entityInnerHostRect,
-      transformOrigin,
-      alertText,
-      isMeasured,
-    } = this.state;
-
-    // Defaulting the main background before passing it to the styles because it is used for `Beak` too.
-    let defaultColor = color;
-    if (!defaultColor && theme) {
-      defaultColor = theme.semanticColors.primaryButtonBackground;
-    }
-
-    const classNames = getClassNames(styles, {
-      theme,
-      beaconColorOne,
-      beaconColorTwo,
-      className,
-      isCollapsed,
-      isBeaconAnimating,
-      isMeasuring,
-      color: defaultColor,
-      transformOrigin,
-      isMeasured,
-      entityHostHeight: `${entityInnerHostRect.height}px`,
-      entityHostWidth: `${entityInnerHostRect.width}px`,
-      width: `${COACHMARK_WIDTH}px`,
-      height: `${COACHMARK_HEIGHT}px`,
-      delayBeforeCoachmarkAnimation: `${delayBeforeCoachmarkAnimation}ms`,
-    });
-
-    const finalHeight: number = isCollapsed ? COACHMARK_HEIGHT : entityInnerHostRect.height;
-
-    return (
-      <PositioningContainer
-        target={target}
-        offsetFromTarget={BEAK_HEIGHT}
-        componentRef={this._positioningContainer}
-        finalHeight={finalHeight}
-        onPositioned={this._onPositioned}
-        bounds={this._getBounds()}
-        {...positioningContainerProps}
-      >
-        <div className={classNames.root}>
-          {ariaAlertText && (
-            <div
-              className={classNames.ariaContainer}
-              role="alert"
-              ref={this._ariaAlertContainer}
-              aria-hidden={!isCollapsed}
-            >
-              {alertText}
-            </div>
-          )}
-          <div className={classNames.pulsingBeacon} />
-          <div className={classNames.translateAnimationContainer} ref={this._translateAnimationContainer}>
-            <div className={classNames.scaleAnimationLayer}>
-              <div className={classNames.rotateAnimationLayer}>
-                {this._positioningContainer.current && (isCollapsed || persistentBeak) && (
-                  <Beak
-                    left={beakLeft}
-                    top={beakTop}
-                    right={beakRight}
-                    bottom={beakBottom}
-                    direction={this._beakDirection}
-                    color={defaultColor}
-                  />
-                )}
-                <div
-                  className={classNames.entityHost}
-                  ref={this._entityHost}
-                  tabIndex={-1}
-                  data-is-focusable={true}
-                  role="dialog"
-                  aria-labelledby={ariaLabelledBy}
-                  aria-describedby={ariaDescribedBy}
-                >
-                  {isCollapsed && [
-                    ariaLabelledBy && (
-                      <p id={ariaLabelledBy} key={0} className={classNames.ariaContainer}>
-                        {ariaLabelledByText}
-                      </p>
-                    ),
-                    ariaDescribedBy && (
-                      <p id={ariaDescribedBy} key={1} className={classNames.ariaContainer}>
-                        {ariaDescribedByText}
-                      </p>
-                    ),
-                  ]}
-                  <FocusTrapZone isClickableOutsideFocusTrap={true} forceFocusInsideTrap={false}>
-                    <div className={classNames.entityInnerHost} ref={this._entityInnerHostElement}>
-                      <div
-                        className={classNames.childrenContainer}
-                        ref={this._childrenContainer}
-                        aria-hidden={isCollapsed}
-                      >
-                        {children}
-                      </div>
-                    </div>
-                  </FocusTrapZone>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </PositioningContainer>
-    );
-  }
-
-  public UNSAFE_componentWillReceiveProps(newProps: ICoachmarkProps): void {
-    if (this.props.isCollapsed && !newProps.isCollapsed) {
-      // The coachmark is about to open
-      this._openCoachmark();
-    }
-  }
-
-  public shouldComponentUpdate(newProps: ICoachmarkProps, newState: ICoachmarkState): boolean {
-    return !shallowCompare(newProps, this.props) || !shallowCompare(newState, this.state);
-  }
-
-  public componentDidUpdate(prevProps: ICoachmarkProps, prevState: ICoachmarkState): void {
-    if (
-      prevState.targetAlignment !== this.state.targetAlignment ||
-      prevState.targetPosition !== this.state.targetPosition
-    ) {
-      this._setBeakPosition();
-    }
-    if (prevProps.preventDismissOnLostFocus !== this.props.preventDismissOnLostFocus) {
-      this._addListeners();
-    }
-  }
-
-  public componentDidMount(): void {
-    this._async.requestAnimationFrame((): void => {
-      if (
-        this._entityInnerHostElement.current &&
-        this.state.entityInnerHostRect.width + this.state.entityInnerHostRect.width === 0
-      ) {
-        this.setState({
-          isMeasuring: false,
-          entityInnerHostRect: {
-            width: this._entityInnerHostElement.current.offsetWidth,
-            height: this._entityInnerHostElement.current.offsetHeight,
-          },
-          isMeasured: true,
-        });
-        this._setBeakPosition();
-        this.forceUpdate();
-      }
-
-      this._addListeners();
-
-      // We don't want to the user to immediately trigger the Coachmark when it's opened
-      this._async.setTimeout(() => {
-        this._addProximityHandler(this.props.mouseProximityOffset);
-      }, this.props.delayBeforeMouseOpen!);
-
-      // Need to add setTimeout to have narrator read change in alert container
-      if (this.props.ariaAlertText) {
-        this._async.setTimeout(() => {
-          if (this.props.ariaAlertText && this._ariaAlertContainer.current) {
-            this.setState({
-              alertText: this.props.ariaAlertText,
-            });
-          }
-        }, 0);
-      }
-
-      if (!this.props.preventFocusOnMount) {
-        this._async.setTimeout(() => {
-          if (this._entityHost.current) {
-            this._entityHost.current.focus();
-          }
-        }, 1000);
-      }
-    });
-  }
-
-  public componentWillUnmount(): void {
-    this._async.dispose();
-    this._events.dispose();
-  }
-
-  public dismiss = (ev?: Event | React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>): void => {
-    const { onDismiss } = this.props;
-
-    if (onDismiss) {
-      onDismiss(ev);
-    }
-  };
-
-  private _addListeners(): void {
-    const { preventDismissOnLostFocus } = this.props;
-    const currentDoc: Document = getDocument()!;
-
-    this._events.off();
-
-    if (currentDoc) {
-      this._events.on(currentDoc, 'keydown', this._onKeyDown, true);
-
-      if (!preventDismissOnLostFocus) {
-        this._events.on(currentDoc, 'click', this._dismissOnLostFocus, true);
-        this._events.on(currentDoc, 'focus', this._dismissOnLostFocus, true);
-      }
-    }
-  }
-
-  private _dismissOnLostFocus(ev: Event) {
-    const clickTarget = ev.target as HTMLElement;
-    const clickedOutsideCallout =
-      this._translateAnimationContainer.current &&
-      !elementContains(this._translateAnimationContainer.current, clickTarget);
-    const { target } = this.props;
-
-    if (clickedOutsideCallout && clickTarget !== target && !elementContains(target as HTMLElement, clickTarget)) {
-      this.dismiss(ev);
-    }
-  }
-
-  private _onKeyDown = (e: any): void => {
-    // Open coachmark if user presses ALT + C (arbitrary keypress for now)
-    if (
-      (e.altKey && e.which === KeyCodes.c) ||
-      (e.which === KeyCodes.enter &&
-        this._translateAnimationContainer.current &&
-        this._translateAnimationContainer.current.contains(e.target))
-    ) {
-      this._onFocusHandler();
-    }
-  };
-
-  private _onFocusHandler = (): void => {
-    if (this.state.isCollapsed) {
-      this._openCoachmark();
-    }
-  };
-
-  private _onPositioned = (positionData: IPositionedData): void => {
-    this._async.requestAnimationFrame((): void => {
-      this.setState({
-        targetAlignment: positionData.alignmentEdge,
-        targetPosition: positionData.targetEdge,
-      });
-    });
-  };
-
-  private _getBounds(): IRectangle | undefined {
-    const { isPositionForced, positioningContainerProps } = this.props;
-    if (isPositionForced) {
-      // If directionalHint direction is the top or bottom auto edge, then we want to set the left/right bounds
-      // to the window x-axis to have auto positioning work correctly.
-      if (
-        positioningContainerProps &&
-        (positioningContainerProps.directionalHint === DirectionalHint.topAutoEdge ||
-          positioningContainerProps.directionalHint === DirectionalHint.bottomAutoEdge)
-      ) {
-        return {
-          left: 0,
-          top: -Infinity,
-          bottom: Infinity,
-          right: window.innerWidth,
-          width: window.innerWidth,
-          height: Infinity,
-        };
-      } else {
-        return {
-          left: -Infinity,
-          top: -Infinity,
-          bottom: Infinity,
-          right: Infinity,
-          width: Infinity,
-          height: Infinity,
-        };
-      }
-    } else {
-      return undefined;
-    }
-  }
-
-  private _setBeakPosition = (): void => {
-    let beakLeft;
-    let beakTop;
-    let beakRight;
-    let beakBottom;
-    let transformOriginX;
-    let transformOriginY;
-
-    const { targetAlignment } = this.state;
     const distanceAdjustment = '3px'; // Adjustment distance for Beak to shift towards Coachmark bubble.
 
-    switch (this._beakDirection) {
+    switch (beakDirection) {
       // If Beak is pointing Up or Down
       case RectangleEdge.top:
       case RectangleEdge.bottom:
         // If there is no target alignment, then beak is X-axis centered in callout
         if (!targetAlignment) {
-          beakLeft = `calc(50% - ${BEAK_WIDTH / 2}px)`;
+          beakPosition.left = `calc(50% - ${BEAK_WIDTH / 2}px)`;
           transformOriginX = 'center';
         } else {
           // Beak is aligned to the left of target
           if (targetAlignment === RectangleEdge.left) {
-            beakLeft = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
+            beakPosition.left = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
             transformOriginX = 'left';
           } else {
             // Beak is aligned to the right of target
-            beakRight = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
+            beakPosition.right = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
             transformOriginX = 'right';
           }
         }
 
-        if (this._beakDirection === RectangleEdge.top) {
-          beakTop = distanceAdjustment;
+        if (beakDirection === RectangleEdge.top) {
+          beakPosition.top = distanceAdjustment;
           transformOriginY = 'top';
         } else {
-          beakBottom = distanceAdjustment;
+          beakPosition.bottom = distanceAdjustment;
           transformOriginY = 'bottom';
         }
         break;
@@ -536,134 +163,410 @@ export class CoachmarkBase extends React.Component<ICoachmarkProps, ICoachmarkSt
       case RectangleEdge.right:
         // If there is no target alignment, then beak is Y-axis centered in callout
         if (!targetAlignment) {
-          beakTop = `calc(50% - ${BEAK_WIDTH / 2}px)`;
+          beakPosition.top = `calc(50% - ${BEAK_WIDTH / 2}px)`;
           transformOriginY = `center`;
         } else {
           // Beak is aligned to the top of target
           if (targetAlignment === RectangleEdge.top) {
-            beakTop = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
+            beakPosition.top = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
             transformOriginY = `top`;
           } else {
             // Beak is aligned to the bottom of target
-            beakBottom = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
+            beakPosition.bottom = `${COACHMARK_WIDTH / 2 - BEAK_WIDTH / 2}px`;
             transformOriginY = `bottom`;
           }
         }
 
-        if (this._beakDirection === RectangleEdge.left) {
-          if (getRTL(this.props.theme)) {
-            beakRight = distanceAdjustment;
+        if (beakDirection === RectangleEdge.left) {
+          if (isRTL) {
+            beakPosition.right = distanceAdjustment;
           } else {
-            beakLeft = distanceAdjustment;
+            beakPosition.left = distanceAdjustment;
           }
           transformOriginX = 'left';
         } else {
-          if (getRTL(this.props.theme)) {
-            beakLeft = distanceAdjustment;
+          if (isRTL) {
+            beakPosition.left = distanceAdjustment;
           } else {
-            beakRight = distanceAdjustment;
+            beakPosition.right = distanceAdjustment;
           }
           transformOriginX = 'right';
         }
         break;
     }
 
-    this.setState({
-      beakLeft: beakLeft,
-      beakRight: beakRight,
-      beakBottom: beakBottom,
-      beakTop: beakTop,
-      transformOrigin: `${transformOriginX} ${transformOriginY}`,
-    });
-  };
+    return [beakPosition as Readonly<BeakPosition>, `${transformOriginX} ${transformOriginY}`] as const;
+  }, [targetAlignment, targetPosition, isRTL]);
+}
 
-  private _openCoachmark = (): void => {
-    this.setState({
-      isCollapsed: false,
-    });
+function useListeners(
+  props: ICoachmarkProps,
+  translateAnimationContainer: React.RefObject<HTMLDivElement>,
+  openCoachmark: () => void,
+) {
+  const document = getDocument()?.documentElement;
 
-    if (this.props.onAnimationOpenStart) {
-      this.props.onAnimationOpenStart();
+  useOnEvent(
+    document,
+    'keydown',
+    (e: KeyboardEvent) => {
+      // Open coachmark if user presses ALT + C (arbitrary keypress for now)
+      if (
+        // eslint-disable-next-line deprecation/deprecation
+        (e.altKey && e.which === KeyCodes.c) ||
+        // eslint-disable-next-line deprecation/deprecation
+        (e.which === KeyCodes.enter && translateAnimationContainer.current?.contains?.(e.target as Node))
+      ) {
+        openCoachmark();
+      }
+    },
+    true,
+  );
+
+  const dismissOnLostFocus = (ev: Event) => {
+    if (props.preventDismissOnLostFocus) {
+      const clickTarget = ev.target as HTMLElement;
+      const clickedOutsideCallout =
+        translateAnimationContainer.current && !elementContains(translateAnimationContainer.current, clickTarget);
+
+      const { target } = props;
+
+      if (clickedOutsideCallout && clickTarget !== target && !elementContains(target as HTMLElement, clickTarget)) {
+        props.onDismiss?.(ev);
+      }
     }
-
-    this._entityInnerHostElement.current &&
-      this._entityInnerHostElement.current.addEventListener('transitionend', (): void => {
-        // Need setTimeout to trigger narrator
-        this._async.setTimeout(() => {
-          if (this._entityInnerHostElement.current) {
-            focusFirstChild(this._entityInnerHostElement.current);
-          }
-        }, 1000);
-
-        if (this.props.onAnimationOpenEnd) {
-          this.props.onAnimationOpenEnd();
-        }
-      });
   };
 
-  private _addProximityHandler(mouseProximityOffset: number = 0): void {
-    /**
-     * An array of cached ids returned when setTimeout runs during
-     * the window resize event trigger.
-     */
-    const timeoutIds: number[] = [];
+  useOnEvent(document, 'click', dismissOnLostFocus, true);
+  useOnEvent(document, 'focus', dismissOnLostFocus, true);
+}
 
-    // Take the initial measure out of the initial render to prevent
-    // an unnecessary render.
-    this._async.setTimeout(() => {
-      this._setTargetElementRect();
+function useProximityHandlers(
+  props: ICoachmarkProps,
+  translateAnimationContainer: React.RefObject<HTMLDivElement>,
+  openCoachmark: () => void,
+) {
+  const { setTimeout, clearTimeout } = useSetTimeout();
 
-      // When the window resizes we want to async
-      // get the bounding client rectangle.
-      // Every time the event is triggered we want to
-      // setTimeout and then clear any previous instances
-      // of setTimeout.
-      this._events.on(window, 'resize', (): void => {
-        timeoutIds.forEach((value: number): void => {
-          clearInterval(value);
+  /** The target element the mouse would be in proximity to */
+  const targetElementRect = React.useRef<DOMRect>();
+
+  React.useEffect(() => {
+    const setTargetElementRect = (): void => {
+      if (translateAnimationContainer.current) {
+        targetElementRect.current = translateAnimationContainer.current.getBoundingClientRect();
+      }
+    };
+
+    const events = new EventGroup({});
+
+    // We don't want the user to immediately trigger the Coachmark when it's opened
+    setTimeout(() => {
+      const { mouseProximityOffset = 0 } = props;
+
+      /** Cached ids returned when setTimeout runs during the window resize event trigger. */
+      const timeoutIds: number[] = [];
+
+      // Take the initial measure out of the initial render to prevent an unnecessary render.
+      setTimeout(() => {
+        setTargetElementRect();
+
+        // When the window resizes we want to async get the bounding client rectangle.
+        // Every time the event is triggered we want to setTimeout and then clear any previous
+        // instances of setTimeout.
+        events.on(window, 'resize', (): void => {
+          timeoutIds.forEach((value: number): void => {
+            clearTimeout(value);
+          });
+          timeoutIds.splice(0, timeoutIds.length); // clear array
+
+          timeoutIds.push(
+            setTimeout((): void => {
+              setTargetElementRect();
+            }, 100),
+          );
         });
+      }, 10);
 
-        timeoutIds.push(
-          this._async.setTimeout((): void => {
-            this._setTargetElementRect();
-          }, 100),
-        );
-      });
-    }, 10);
-
-    // Every time the document's mouse move is triggered
-    // we want to check if inside of an element and
-    // set the state with the result.
-    this._events.on(document, 'mousemove', (e: MouseEvent) => {
-      if (this.state.isCollapsed) {
+      // Every time the document's mouse move is triggered, we want to check if inside of an element
+      // and set the state with the result.
+      events.on(document, 'mousemove', (e: MouseEvent) => {
         const mouseY = e.clientY;
         const mouseX = e.clientX;
-        this._setTargetElementRect();
-        const isMouseInProximity = this._isInsideElement(mouseX, mouseY, mouseProximityOffset);
+        setTargetElementRect();
 
-        if (isMouseInProximity !== this.state.isMouseInProximity) {
-          this._openCoachmark();
+        if (isInsideElement(targetElementRect.current!, mouseX, mouseY, mouseProximityOffset)) {
+          openCoachmark();
         }
-      }
 
-      if (this.props.onMouseMove) {
-        this.props.onMouseMove(e);
+        props.onMouseMove?.(e);
+      });
+    }, props.delayBeforeMouseOpen!);
+
+    return () => events.dispose();
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- should only run on mount
+  }, []);
+}
+
+function useComponentRef(props: ICoachmarkProps) {
+  const { onDismiss } = props;
+  React.useImperativeHandle(
+    props.componentRef,
+    (ev?: Event | React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>) => ({
+      dismiss() {
+        onDismiss?.(ev);
+      },
+    }),
+    [onDismiss],
+  );
+}
+
+function useAriaAlert({ ariaAlertText }: ICoachmarkProps) {
+  const async = useAsync();
+
+  /** ARIA alert text to read aloud with Narrator once the Coachmark is mounted */
+  const [alertText, setAlertText] = React.useState<string | undefined>();
+
+  React.useEffect(() => {
+    // Need to add RAF to have narrator read change in alert container
+    async.requestAnimationFrame(() => {
+      setAlertText(ariaAlertText);
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- should only run on mount
+  }, []);
+
+  return alertText;
+}
+
+function useAutoFocus({ preventFocusOnMount }: ICoachmarkProps) {
+  const { setTimeout } = useSetTimeout();
+
+  /**
+   * The cached HTMLElement reference to the Entity Inner Host
+   * element.
+   */
+  const entityHost = React.useRef<HTMLDivElement | null>(null);
+
+  React.useEffect(() => {
+    if (!preventFocusOnMount) {
+      setTimeout(() => entityHost.current?.focus(), 1000);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- should only run on mount
+  }, []);
+
+  return entityHost;
+}
+
+function useEntityHostMeasurements(props: ICoachmarkProps, entityInnerHostElementRef: React.RefObject<HTMLDivElement>) {
+  /** Is the teaching bubble currently retrieving the original dimensions of the hosted entity. */
+  const [isMeasuring, setIsMeasuring] = React.useState<boolean>(!!props.isCollapsed);
+  /** Cached width and height of _entityInnerHostElement */
+  const [entityInnerHostRect, setEntityInnerHostRect] = React.useState<IEntityRect>(
+    props.isCollapsed ? { width: 0, height: 0 } : {},
+  );
+  const async = useAsync();
+
+  React.useEffect(() => {
+    async.requestAnimationFrame(() => {
+      if (entityInnerHostElementRef.current) {
+        setEntityInnerHostRect({
+          width: entityInnerHostElementRef.current.offsetWidth,
+          height: entityInnerHostElementRef.current.offsetHeight,
+        });
+        setIsMeasuring(false);
       }
     });
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- should only run on mount
+  }, []);
+
+  return [isMeasuring, entityInnerHostRect] as const;
+}
+
+function useDeprecationWarning(props: ICoachmarkProps) {
+  if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks -- build-time conditional
+    useWarnings({
+      name: COMPONENT_NAME,
+      props,
+      deprecations: {
+        teachingBubbleRef: undefined,
+        collapsed: 'isCollapsed',
+        beakWidth: undefined,
+        beakHeight: undefined,
+        width: undefined,
+        height: undefined,
+      },
+    });
+  }
+}
+
+const COMPONENT_NAME = 'CoachmarkBase';
+
+export const CoachmarkBase: React.FunctionComponent<ICoachmarkProps> = React.forwardRef<
+  HTMLDivElement,
+  ICoachmarkProps
+>((propsWithoutDefaults, forwardedRef) => {
+  const props = getPropsWithDefaults(DEFAULT_PROPS, propsWithoutDefaults);
+
+  const entityInnerHostElementRef = React.useRef<HTMLDivElement | null>(null);
+  const translateAnimationContainer = React.useRef<HTMLDivElement | null>(null);
+
+  const [targetAlignment, targetPosition, onPositioned] = usePositionedData();
+  const [isCollapsed, openCoachmark] = useCollapsedState(props, entityInnerHostElementRef);
+  const [beakPositioningProps, transformOrigin] = useBeakPosition(props, targetAlignment, targetPosition);
+  const [isMeasuring, entityInnerHostRect] = useEntityHostMeasurements(props, entityInnerHostElementRef);
+  const alertText = useAriaAlert(props);
+  const entityHost = useAutoFocus(props);
+
+  useListeners(props, translateAnimationContainer, openCoachmark);
+  useComponentRef(props);
+  useProximityHandlers(props, translateAnimationContainer, openCoachmark);
+  useDeprecationWarning(props);
+
+  const {
+    beaconColorOne,
+    beaconColorTwo,
+    children,
+    target,
+    color,
+    positioningContainerProps,
+    ariaDescribedBy,
+    ariaDescribedByText,
+    ariaLabelledBy,
+    ariaLabelledByText,
+    ariaAlertText,
+    delayBeforeCoachmarkAnimation,
+    styles,
+    theme,
+    className,
+    persistentBeak,
+  } = props;
+
+  // Defaulting the main background before passing it to the styles because it is used for `Beak` too.
+  let defaultColor = color;
+  if (!defaultColor && theme) {
+    defaultColor = theme.semanticColors.primaryButtonBackground;
   }
 
-  private _setTargetElementRect(): void {
-    if (this._translateAnimationContainer && this._translateAnimationContainer.current) {
-      this._targetElementRect = this._translateAnimationContainer!.current!.getBoundingClientRect();
+  const classNames = getClassNames(styles, {
+    theme,
+    beaconColorOne,
+    beaconColorTwo,
+    className,
+    isCollapsed,
+    isMeasuring,
+    color: defaultColor,
+    transformOrigin,
+    entityHostHeight: entityInnerHostRect.height === undefined ? undefined : `${entityInnerHostRect.height}px`,
+    entityHostWidth: entityInnerHostRect.width === undefined ? undefined : `${entityInnerHostRect.width}px`,
+    width: `${COACHMARK_WIDTH}px`,
+    height: `${COACHMARK_HEIGHT}px`,
+    delayBeforeCoachmarkAnimation: `${delayBeforeCoachmarkAnimation}ms`,
+  });
+
+  const finalHeight: number | undefined = isCollapsed ? COACHMARK_HEIGHT : entityInnerHostRect.height;
+
+  return (
+    <PositioningContainer
+      target={target}
+      offsetFromTarget={BEAK_HEIGHT}
+      finalHeight={finalHeight}
+      ref={forwardedRef}
+      onPositioned={onPositioned}
+      bounds={getBounds(props)}
+      {...positioningContainerProps}
+    >
+      <div className={classNames.root}>
+        {ariaAlertText && (
+          <div className={classNames.ariaContainer} role="alert" aria-hidden={!isCollapsed}>
+            {alertText}
+          </div>
+        )}
+        <div className={classNames.pulsingBeacon} />
+        <div className={classNames.translateAnimationContainer} ref={translateAnimationContainer}>
+          <div className={classNames.scaleAnimationLayer}>
+            <div className={classNames.rotateAnimationLayer}>
+              {(isCollapsed || persistentBeak) && <Beak {...beakPositioningProps} color={defaultColor} />}
+              <div
+                className={classNames.entityHost}
+                ref={entityHost}
+                tabIndex={-1}
+                data-is-focusable={true}
+                role="dialog"
+                aria-labelledby={ariaLabelledBy}
+                aria-describedby={ariaDescribedBy}
+              >
+                {isCollapsed && [
+                  ariaLabelledBy && (
+                    <p id={ariaLabelledBy} key={0} className={classNames.ariaContainer}>
+                      {ariaLabelledByText}
+                    </p>
+                  ),
+                  ariaDescribedBy && (
+                    <p id={ariaDescribedBy} key={1} className={classNames.ariaContainer}>
+                      {ariaDescribedByText}
+                    </p>
+                  ),
+                ]}
+                <FocusTrapZone isClickableOutsideFocusTrap={true} forceFocusInsideTrap={false}>
+                  <div className={classNames.entityInnerHost} ref={entityInnerHostElementRef}>
+                    <div className={classNames.childrenContainer} aria-hidden={isCollapsed}>
+                      {children}
+                    </div>
+                  </div>
+                </FocusTrapZone>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </PositioningContainer>
+  );
+});
+CoachmarkBase.displayName = COMPONENT_NAME;
+
+function getBounds({ isPositionForced, positioningContainerProps }: ICoachmarkProps): IRectangle | undefined {
+  if (isPositionForced) {
+    // If directionalHint direction is the top or bottom auto edge, then we want to set the left/right bounds
+    // to the window x-axis to have auto positioning work correctly.
+    if (
+      positioningContainerProps &&
+      (positioningContainerProps.directionalHint === DirectionalHint.topAutoEdge ||
+        positioningContainerProps.directionalHint === DirectionalHint.bottomAutoEdge)
+    ) {
+      return {
+        left: 0,
+        top: -Infinity,
+        bottom: Infinity,
+        right: window.innerWidth,
+        width: window.innerWidth,
+        height: Infinity,
+      };
+    } else {
+      return {
+        left: -Infinity,
+        top: -Infinity,
+        bottom: Infinity,
+        right: Infinity,
+        width: Infinity,
+        height: Infinity,
+      };
     }
+  } else {
+    return undefined;
   }
+}
 
-  private _isInsideElement(mouseX: number, mouseY: number, mouseProximityOffset: number = 0): boolean {
-    return (
-      mouseX > this._targetElementRect.left - mouseProximityOffset &&
-      mouseX < this._targetElementRect.left + this._targetElementRect.width + mouseProximityOffset &&
-      mouseY > this._targetElementRect.top - mouseProximityOffset &&
-      mouseY < this._targetElementRect.top + this._targetElementRect.height + mouseProximityOffset
-    );
-  }
+function isInsideElement(
+  targetElementRect: ClientRect,
+  mouseX: number,
+  mouseY: number,
+  mouseProximityOffset: number = 0,
+): boolean {
+  return (
+    mouseX > targetElementRect.left - mouseProximityOffset &&
+    mouseX < targetElementRect.left + targetElementRect.width + mouseProximityOffset &&
+    mouseY > targetElementRect.top - mouseProximityOffset &&
+    mouseY < targetElementRect.top + targetElementRect.height + mouseProximityOffset
+  );
 }
