--- packages/react-internal/src/components/Coachmark/Coachmark.test.ts
@@ -1,9 +1,0 @@
-import { Coachmark } from './Coachmark';
-import { isConformant } from '../../common/isConformant';
-
-describe('Coachmark', () => {
-  isConformant({
-    Component: Coachmark,
-    displayName: 'Coachmark',
-  });
-});
