--- packages/react-next/src/components/Coachmark/Coachmark.ts
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import { styled } from '../../Utilities';
-import { ICoachmarkProps, ICoachmarkStyleProps, ICoachmarkStyles } from './Coachmark.types';
-import { getStyles } from './Coachmark.styles';
-import { CoachmarkBase } from './Coachmark.base';
-
-export const Coachmark: React.FunctionComponent<ICoachmarkProps> = styled<
-  ICoachmarkProps,
-  ICoachmarkStyleProps,
-  ICoachmarkStyles
->(CoachmarkBase, getStyles, undefined, {
-  scope: 'Coachmark',
-});
