--- packages/react-next/src/components/Coachmark/PositioningContainer/index.ts
@@ -1,2 +1,0 @@
-export * from './PositioningContainer';
-export * from './PositioningContainer.types';
