--- packages/react-next/src/components/Coachmark/index.ts
@@ -1,3 +1,0 @@
-export * from './Coachmark';
-export * from './Coachmark.base';
-export * from './Coachmark.types';
